Made changes to move to Postgre

+++ src/main/java/org/dbhaskaran/covid19/repos/ICovid.java
@@ -3,6 +3,6 @@ package org.dbhaskaran.covid19.repos;
 import org.dbhaskaran.covid19.entities.Covid;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ICovid extends JpaRepository<Covid, Integer> {
+public interface ICovid extends JpaRepository<Covid, Long> {
 
 }
